Move bodhi.server.schemas.splitter to bodhi.server.util.splitter.
This commit solves a circular import problem. bodhi.server.schemas
imports bodhi.server.validators, and bodhi.server.validators was
importing bodhi.server.schemas to use its splitter() function. This
commit moves that function to bodhi.server.util so they can each
import it from there and avoid the circular imports.

fixes #1186
